Remove Jade, Node, and Browserify
This removes the dependencies on Jade, Node, and Browserify and makes it
a simple vanilla js sample.
