--- .gitignore
@@ -1,1 +1,0 @@
-js-sample/public/app.js
