+++ js-sample/.jshintrc
@@ -1,0 +1,21 @@
+{
+  "browserify": true,
+  "globalstrict": true,
+  "node": true,
+  "browser": true,
+  "esnext": true,
+  "strict": false,
+  "camelcase": true,
+  "eqeqeq": true,
+  "eqnull": true,
+  "expr": true,
+  "indent": 2,
+  "latedef": true,
+  "maxlen": 100,
+  "newcap": true,
+  "quotmark": "single",
+  "smarttabs": true,
+  "trailing": true,
+  "undef": true,
+  "unused": true
+}
