+++ js-sample/index.html
@@ -1,0 +1,16 @@
+<html>
+  <head>
+    <script type="text/javascript" src="TB.js"></script>
+    <script type="text/javascript" src="app.js"></script>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link href="assets/styles.css"  rel="stylesheet">
+  </head>
+  <body>
+    <h1>OpenTok <small>Network test</small>
+    <div id="status_container" class="prettyBox">
+      <h2>Testing stream capabilities</h2>
+      <p>Acquiring camera</p>
+      <img src="assets/spinner.gif">
+    </div>
+  </body>
+</html>
