--- js-sample/src/js/lib/quality-test.js
@@ -1,32 +1,0 @@
-var BandwidthCalculator = require('./bandwidth-calculator.js');
-
-module.exports = function(config, callback) {
-  var startMs = new Date().getTime(),
-      testTimeout,
-      currentStats;
-
-  var bandwidthCalculator = BandwidthCalculator({
-    subscriber: config.subscriber
-  });
-
-  var cleanupAndReport = function() {
-    currentStats.elapsedTimeMs = new Date().getTime() - startMs;
-    callback(undefined, currentStats);
-
-    window.clearTimeout(testTimeout);
-    bandwidthCalculator.stop();
-
-    callback = function() {};
-  };
-
-  // bail out of the test after 30 seconds
-  window.setTimeout(cleanupAndReport, config.timeout);
-
-  bandwidthCalculator.start(function(stats) {
-    console.log(stats);
-
-    // you could do something smart here like determine if the bandwidth is
-    // stable or acceptable and exit early
-    currentStats = stats;
-  });
-};
