--- js-sample/src/js/lib/sample-window-size.js
@@ -1,8 +1,0 @@
-var max = require('lodash.max'),
-    min = require('lodash.min'),
-    pluck = require('lodash.pluck');
-
-module.exports = function(samples) {
-  var times = pluck(samples, 'timestamp');
-  return (max(times) - min(times)) / 1000;
-};
